# Fix of issue #74 and #178 (only interpreter and x86 JIT code gen)