… lib/common/common/NumberUtilities.h
… test/Math/round.js