Checks are added to solve issue #74 and #178.