@mcfi 2016-01-28
This PR replaces #186 due to its enhanced performance by avoiding double ```ROUNDSD``` instructions. The benchmark performance results are below,

```
x86release
Baseline file: perfbaseoctane.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x86_release/ch.exe
Iterations: 10
High Precision Date not used!
Running octane...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
box2d                    36314.3 +-1.7% 36194.2 +-1.2%  -120.1   -0.3%
code-load                20645.0 +-0.4% 20659.3 +-0.3%    14.3    0.1%
crypto                   30282.2 +-0.9% 30124.1 +-1.2%  -158.1   -0.5%
deltablue                26321.6 +-1.5% 26218.7 +-1.4%  -102.9   -0.4%
earley-boyer             49443.2 +-0.7% 49451.7 +-0.8%     8.4    0.0%
gbemu                    59064.7 +-2.5% 58717.7 +-2.0%  -347.0   -0.6%
mandreel                 29239.4 +-0.6% 29197.8 +-0.4%   -41.7   -0.1%
mandreelLatency          101983.8 +-1.2%        101993.1 +-1.5%    9.3    0.0%
navier-stokes            37433.3 +-0.1% 37425.2 +-0.2%    -8.1   -0.0%
pdfjs                    26300.1 +-0.8% 26409.1 +-1.1%   109.0    0.4%
raytrace                 44712.9 +-2.4% 44083.7 +-4.2%  -629.2   -1.4%
regexp                   4623.8 +-0.5%  4667.8 +-0.6%     44.0    1.0%
richards                 28531.9 +-1.0% 28405.6 +-1.0%  -126.3   -0.4%
splay                    20503.0 +-18.5%        21396.8 +-10.0%  893.8    4.4%
splayLatency             18612.3 +-63.5%        13944.4 +-21.1% -4667.9 -25.1%
typescript               33934.4 +-1.6% 33734.1 +-2.3%  -200.3   -0.6%
zlib                     56915.1 +-0.9% 56880.6 +-0.5%   -34.6   -0.1%
---------------------------------------------------------------------------
MEAN SCORE               30977.5 +-1.4% 30477.5 +-1.3%  -500.0   -1.6%

Error in opening file '/?' : The filename, directory name, or volume label syntax is incorrect.

Warning: Persistent Dynamic profile is not supported!
Baseline file: perfbasekraken.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x86_release/ch.exe
Iterations: 10
High Precision Date not used!
Running kraken...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
ai-astar                  216.9 +-0.5%   216.4 +-0.2%     -0.4   -0.2%
audio-beat-detection       74.2 +-2.8%    73.4 +-0.7%     -0.8   -1.0%
audio-dft                 159.3 +-0.6%   164.7 +-5.0%      5.3    3.3%
audio-fft                  53.4 +-5.6%    52.4 +-1.8%     -1.0   -1.9%
audio-oscillator           55.9 +-0.6%    62.1 +-0.5%      6.2   11.1% <-CHECK
imaging-darkroom          123.8 +-0.3%   124.3 +-0.4%      0.6    0.4%
imaging-desaturate         49.0 +-0.0%    49.1 +-0.6%      0.1    0.2%
imaging-gaussian-blur     124.7 +-0.8%   124.3 +-1.1%     -0.3   -0.3%
json-parse-financial       43.7 +-12.5%   42.0 +-6.7%     -1.7   -3.8%
json-stringify-tinderbox   37.0 +-0.0%    37.0 +-0.0%      0.0    0.0%
stanford-crypto-aes       108.4 +-5.1%   108.1 +-1.9%     -0.3   -0.3%
stanford-crypto-ccm        82.3 +-2.8%    84.6 +-4.3%      2.2    2.7%
stanford-crypto-pbkdf2    233.7 +-2.0%   235.3 +-3.1%      1.7    0.7%
stanford-crypto-sha256-i   53.2 +-1.5%    55.6 +-7.1%      2.3    4.4%
---------------------------------------------------------------------------
TOTAL TIME               1415.6 +-1.9%  1429.4 +-2.3%     13.9    1.0%

Error in opening file '/?' : The filename, directory name, or volume label syntax is incorrect.

Warning: Persistent Dynamic profile is not supported!
Baseline file: perfbasesunspider.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x86_release/ch.exe
Iterations: 35
High Precision Date not used!
Running sunspider...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
3d-cube                    10.0 +-2.5%    10.0 +-3.1%      0.0    0.3%
3d-morph                    2.0 +-0.0%     2.0 +-0.0%      0.0    0.0%
3d-raytrace                 9.9 +-2.5%    10.0 +-0.0%      0.1    0.6%
access-binary-trees         3.3 +-13.4%    3.1 +-8.0%     -0.2   -5.9%
access-fannkuch             6.1 +-6.4%     6.2 +-10.6%     0.1    2.1%
access-nbody                5.7 +-14.0%    5.7 +-14.2%    -0.0   -0.6%
access-nsieve               2.0 +-0.0%     2.0 +-0.0%      0.0    0.0%
bitops-3bit-bits-in-byte    1.0 +-0.0%     1.0 +-0.0%      0.0    0.0%
bitops-bits-in-byte         2.2 +-18.8%    2.2 +-17.0%    -0.1   -2.9%
bitops-bitwise-and          2.0 +-0.0%     2.0 +-0.0%      0.0    0.0%
bitops-nsieve-bits          4.0 +-0.0%     4.0 +-0.0%      0.0    0.0%
controlflow-recursive       5.4 +-10.2%    5.4 +-9.1%      0.0    0.6%
crypto-aes                  8.2 +-7.4%     8.0 +-0.0%     -0.2   -2.7%
crypto-md5                  6.3 +-10.7%    6.0 +-0.0%     -0.3   -4.1%
crypto-sha1                 4.6 +-11.7%    4.8 +-11.5%     0.1    2.8%
date-format-tofte           9.1 +-4.6%     9.4 +-10.0%     0.2    2.5%
date-format-xparb          12.5 +-5.3%    13.1 +-7.8%      0.6    4.6%
math-cordic                 2.2 +-18.0%    2.0 +-0.0%     -0.2   -8.8%
math-partial-sums           7.0 +-0.0%     7.0 +-0.0%      0.0    0.0%
math-spectral-norm          3.0 +-6.0%     3.0 +-6.0%      0.0    0.0%
regexp-dna                  5.0 +-0.0%     5.0 +-0.0%      0.0    0.0%
string-base64               5.0 +-0.0%     5.0 +-0.0%      0.0    0.0%
string-fasta                8.0 +-0.0%     8.0 +-0.0%      0.0    0.0%
string-tagcloud            16.7 +-2.6%    17.1 +-5.6%      0.3    1.9%
string-unpack-code         17.0 +-0.0%    17.1 +-2.3%      0.1    0.6%
string-validate-input       7.4 +-11.2%    7.3 +-8.7%     -0.1   -0.9%
---------------------------------------------------------------------------
TOTAL TIME                165.7 +-4.7%   166.3 +-4.5%      0.6    0.4%

Error in opening file '/?' : The filename, directory name, or volume label syntax is incorrect.

Baseline file: perfbasejetstream.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x86_release/ch.exe
Iterations: 5
High Precision Date not used!
Running jetstream...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
bigfib.cpp                281.4 +-1.1%   279.0 +-0.8%     -2.4   -0.9%
container.cpp             332.6 +-0.9%   332.8 +-0.5%      0.2    0.1%
dry.c                     257.2 +-2.0%   262.0 +-1.9%      4.8    1.9%
float-mm.c                351.0 +-2.9%   350.2 +-1.1%     -0.8   -0.2%
gcc-loops.cpp             360.6 +-3.0%   363.2 +-1.8%      2.6    0.7%
hash-map                  152.8 +-1.5%   151.8 +-2.4%     -1.0   -0.7%
n-body.c                  234.8 +-0.5%   233.0 +-2.0%     -1.8   -0.8%
quicksort.c               211.4 +-1.4%   213.2 +-1.3%      1.8    0.9%
towers.c                  239.4 +-0.5%   238.6 +-0.4%     -0.8   -0.3%
cdjs                      186.6 +-3.2%   180.8 +-6.8%     -5.8   -3.1%
---------------------------------------------------------------------------
MEAN SCORE                252.0 +-1.4%   251.4 +-1.4%     -0.6   -0.3%

x64release
Baseline file: perfbaseoctane.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x64_release/ch.exe
Iterations: 10
High Precision Date not used!
Running octane...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
box2d                    43718.9 +-0.9% 43596.8 +-0.8%  -122.1   -0.3%
code-load                21126.0 +-0.3% 21263.3 +-0.3%   137.3    0.7%
crypto                   31467.9 +-1.5% 31446.0 +-0.9%   -21.9   -0.1%
deltablue                22759.3 +-4.5% 22984.1 +-6.1%   224.8    1.0%
earley-boyer             42838.8 +-1.0% 42681.6 +-0.6%  -157.2   -0.4%
gbemu                    58836.1 +-0.3% 59216.1 +-0.7%   380.0    0.6%
mandreel                 29899.3 +-0.4% 29858.1 +-0.8%   -41.2   -0.1%
mandreelLatency          106324.2 +-0.7%        106591.1 +-1.3%  266.9    0.3%
navier-stokes            37614.2 +-0.7% 37857.2 +-0.2%   243.0    0.6%
pdfjs                    23937.4 +-2.1% 23979.1 +-1.5%    41.7    0.2%
raytrace                 43029.4 +-3.5% 43766.7 +-5.4%   737.2    1.7%
regexp                   5022.4 +-0.6%  5104.3 +-0.7%     81.9    1.6%
richards                 27353.6 +-1.2% 27537.2 +-1.4%   183.7    0.7%
splay                    18305.2 +-12.7%        17924.6 +-8.9%  -380.7   -2.1%
splayLatency             15135.2 +-57.4%        12364.9 +-31.9% -2770.3 -18.3%
typescript               27786.7 +-2.3% 27725.2 +-1.9%   -61.4   -0.2%
zlib                     65853.0 +-1.5% 66403.6 +-0.9%   550.6    0.8%
---------------------------------------------------------------------------
MEAN SCORE               30180.7 +-1.5% 29911.4 +-1.4%  -269.2   -0.9%

Error in opening file '/?' : The filename, directory name, or volume label syntax is incorrect.

Warning: Persistent Dynamic profile is not supported!
Baseline file: perfbasekraken.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x64_release/ch.exe
Iterations: 10
High Precision Date not used!
Running kraken...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
ai-astar                  270.3 +-0.6%   269.8 +-0.5%     -0.6   -0.2%
audio-beat-detection       75.7 +-4.0%    75.1 +-1.5%     -0.6   -0.7%
audio-dft                 161.3 +-2.9%   161.0 +-0.5%     -0.3   -0.2%
audio-fft                  54.7 +-2.4%    54.1 +-0.6%     -0.6   -1.0%
audio-oscillator           57.1 +-6.8%    60.8 +-3.1%      3.7    6.4%
imaging-darkroom          119.8 +-0.3%   119.7 +-0.4%     -0.1   -0.1%
imaging-desaturate         48.3 +-1.0%    48.4 +-1.0%      0.1    0.2%
imaging-gaussian-blur     123.3 +-1.6%   123.2 +-0.6%     -0.1   -0.1%
json-parse-financial       37.7 +-1.3%    38.0 +-0.0%      0.3    0.9%
json-stringify-tinderbox   35.9 +-0.9%    36.7 +-3.4%      0.8    2.2%
stanford-crypto-aes       113.9 +-2.4%   113.2 +-1.0%     -0.7   -0.6%
stanford-crypto-ccm        78.1 +-1.3%    81.7 +-4.0%      3.6    4.6% <-CHECK
stanford-crypto-pbkdf2    125.4 +-0.5%   125.6 +-0.9%      0.1    0.1%
stanford-crypto-sha256-i   46.8 +-20.6%   47.0 +-15.4%     0.2    0.5%
---------------------------------------------------------------------------
TOTAL TIME               1348.3 +-2.4%  1354.2 +-1.6%      5.9    0.4%

Error in opening file '/?' : The filename, directory name, or volume label syntax is incorrect.

Warning: Persistent Dynamic profile is not supported!
Baseline file: perfbasesunspider.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x64_release/ch.exe
Iterations: 35
High Precision Date not used!
Running sunspider...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
3d-cube                     9.2 +-4.3%     9.3 +-4.7%      0.1    0.7%
3d-morph                    2.0 +-0.0%     2.0 +-0.0%      0.0    0.0%
3d-raytrace                10.0 +-1.8%    10.0 +-0.0%     -0.0   -0.3%
access-binary-trees         3.2 +-11.6%    3.2 +-12.4%     0.0    1.0%
access-fannkuch             6.0 +-0.0%     6.4 +-12.4%     0.4    7.0%
access-nbody                4.0 +-0.0%     4.1 +-8.1%      0.1    3.2%
access-nsieve               2.0 +-0.0%     2.0 +-0.0%      0.0    0.0%
bitops-3bit-bits-in-byte    1.0 +-0.0%     1.0 +-0.0%      0.0    0.0%
bitops-bits-in-byte         2.1 +-14.1%    2.2 +-18.0%     0.1    4.6%
bitops-bitwise-and          2.0 +-0.0%     2.0 +-0.0%      0.0    0.0%
bitops-nsieve-bits          2.5 +-20.3%    2.3 +-20.1%    -0.1   -5.3%
controlflow-recursive       5.3 +-8.6%     5.0 +-3.5%     -0.3   -4.9%
crypto-aes                  8.0 +-0.0%     8.0 +-0.0%      0.0    0.0%
crypto-md5                  5.0 +-3.5%     5.0 +-3.5%     -0.0   -0.0%
crypto-sha1                 3.3 +-19.3%    3.1 +-8.0%     -0.3   -7.8%
date-format-tofte           9.8 +-10.6%    9.4 +-5.9%     -0.4   -4.0%
date-format-xparb          15.3 +-10.3%   14.7 +-5.4%     -0.5   -3.6%
math-cordic                 2.1 +-14.1%    2.0 +-0.0%     -0.1   -4.6%
math-partial-sums           6.1 +-6.4%     6.0 +-0.0%     -0.1   -1.6%
math-spectral-norm          2.1 +-15.7%    2.1 +-11.9%    -0.1   -3.0%
regexp-dna                  4.0 +-4.4%     4.0 +-0.0%     -0.0   -0.8%
string-base64               5.0 +-0.0%     5.0 +-0.0%      0.0    0.0%
string-fasta                8.0 +-0.0%     8.1 +-3.0%      0.1    0.8%
string-tagcloud            16.0 +-1.1%    16.0 +-0.0%     -0.0   -0.2%
string-unpack-code         16.8 +-7.5%    16.5 +-3.0%     -0.3   -1.7%
string-validate-input       6.6 +-12.6%    6.6 +-9.2%      0.0    0.0%
---------------------------------------------------------------------------
TOTAL TIME                157.5 +-5.8%   156.0 +-4.1%     -1.4   -0.9%

Error in opening file '/?' : The filename, directory name, or volume label syntax is incorrect.

Baseline file: perfbasejetstream.txt
Binary: C:/Users/niube/Documents/Github/ChakraCore/Build/VcBuild/bin-fix178/x64_release/ch.exe
Iterations: 5
High Precision Date not used!
Running jetstream...

TEST - NATIVE                BASE            TEST         DIFF   RATIO(%)
---------------------------------------------------------------------------
bigfib.cpp                335.6 +-0.8%   330.8 +-1.7%     -4.8   -1.4%
container.cpp             347.0 +-0.4%   346.4 +-0.6%     -0.6   -0.2%
dry.c                     317.4 +-3.7%   313.8 +-3.9%     -3.6   -1.1%
float-mm.c                364.6 +-1.8%   369.6 +-2.4%      5.0    1.4%
gcc-loops.cpp             427.0 +-1.0%   427.2 +-2.1%      0.2    0.0%
hash-map                  130.2 +-3.0%   131.8 +-2.3%      1.6    1.2%
n-body.c                  280.6 +-1.5%   281.6 +-0.8%      1.0    0.4%
quicksort.c               250.6 +-1.3%   254.0 +-0.4%      3.4    1.4%
towers.c                  263.0 +-2.9%   256.2 +-2.8%     -6.8   -2.6%
cdjs                      188.4 +-2.9%   178.0 +-8.1%    -10.4   -5.5%
---------------------------------------------------------------------------
MEAN SCORE                276.5 +-1.6%   274.6 +-1.8%     -1.9   -0.7%
```